feat(sdk): Export AttributeObject and others
- Remove duplicate AttributeObject interface, replacing with type
- Removes unnecessary constructor function
- Exports the type
- Refactor of similar duplication with policy:
   - NanoTDF code `Policy` becomes `PolicyBuilder`
   - The Nano code's `PolicyObject` is removed in favor of the `tdf3.Policy` type
