+++ lib/tdf3/src/models/index.ts
@@ -1,3 +1,4 @@
+export * from './attribute.js';
 export * from './attribute-set.js';
 export * from './encryption-information.js';
 export * from './key-access.js';
