+++ lib/tdf3/src/models/policy.ts
@@ -1,5 +1,5 @@
 import { ConfigurationError } from '../../../src/errors.js';
-import { AttributeObject } from './attribute-set.js';
+import { type AttributeObject } from './attribute.js';
 
 export const CURRENT_VERSION = '1.1.0';
